# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface Category {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  slug: string;
  description?: string;
  descriptionAr?: string;
  descriptionFr?: string;
  icon?: string;
  image?: string;
  productCount?: number;
}

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  slug: string;
  price: number;
  originalPrice?: number | null;
  currency: string;
  image: string;
  categoryId: number;
  categoryName: string;
  categoryNameAr: string;
  categoryNameFr: string;
  inStock: boolean;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  discount?: number | null;
}

export type ProductDetailSpecifications = { [key: string]: string };

export type ProductDetail = Product & {
  description: string;
  descriptionAr: string;
  descriptionFr: string;
  images: string[];
  specifications?: ProductDetailSpecifications;
  stockQuantity: number;
  sku?: string;
  brand?: string;
  tags?: string[];
};

export interface ProductsPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface OrderItem {
  productId: number;
  quantity: number;
  price: number;
  name: string;
}

export interface CreateOrderRequest {
  fullName: string;
  phone: string;
  email?: string | null;
  wilaya: string;
  city: string;
  address?: string | null;
  items: OrderItem[];
  notes?: string | null;
}

export interface CreateInstantOrderRequest {
  fullName: string;
  phone: string;
  wilaya: string;
  city: string;
  address?: string | null;
  productId: number;
  quantity: number;
  notes?: string | null;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  fullName: string;
  phone: string;
  wilaya: string;
  city: string;
  address?: string | null;
  total: number;
  items: OrderItem[];
  createdAt: string;
}

export interface City {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  wilayaCode: string;
}

export interface Region {
  id: number;
  code: string;
  name: string;
  nameAr: string;
  nameFr: string;
  cities: City[];
}

export type GetProductsParams = {
  search?: string;
  categoryId?: number;
  minPrice?: number;
  maxPrice?: number;
  sortBy?: GetProductsSortBy;
  page?: number;
  limit?: number;
  inStock?: boolean;
};

export type GetProductsSortBy =
  (typeof GetProductsSortBy)[keyof typeof GetProductsSortBy];

export const GetProductsSortBy = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  newest: "newest",
  popular: "popular",
} as const;