# lib/api-zod/src/generated/types/city.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface City {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  wilayaCode: string;
}