# lib/api-zod/src/generated/types/createInstantOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInstantOrderRequest {
  fullName: string;
  phone: string;
  wilaya: string;
  city: string;
  address?: string | null;
  productId: number;
  quantity: number;
  notes?: string | null;
}