# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface CreateOrderRequest {
  fullName: string;
  phone: string;
  email?: string | null;
  wilaya: string;
  city: string;
  address?: string | null;
  items: OrderItem[];
  notes?: string | null;
}