# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  slug: string;
  price: number;
  originalPrice?: number | null;
  currency: string;
  image: string;
  categoryId: number;
  categoryName: string;
  categoryNameAr: string;
  categoryNameFr: string;
  inStock: boolean;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  discount?: number | null;
}