# lib/api-zod/src/generated/types/productDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";
import type { ProductDetailSpecifications } from "./productDetailSpecifications";

export type ProductDetail = Product & {
  description: string;
  descriptionAr: string;
  descriptionFr: string;
  images: string[];
  specifications?: ProductDetailSpecifications;
  stockQuantity: number;
  sku?: string;
  brand?: string;
  tags?: string[];
};