# lib/api-zod/src/generated/types/region.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikDz Algerian Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { City } from "./city";

export interface Region {
  id: number;
  code: string;
  name: string;
  nameAr: string;
  nameFr: string;
  cities: City[];
}